优化了一系列空岛战争的细节

+++ src/main/java/cn/nukkit/block/BlockTNT.java
@@ -97,7 +97,7 @@ public class BlockTNT extends BlockSolidMeta {
                 .putList(new ListTag<FloatTag>("Rotation")
                         .add(new FloatTag("", 0))
                         .add(new FloatTag("", 0)))
-                .putShort("Fuse", fuse);
+                .putByte("Fuse", fuse);
         Entity tnt = new EntityPrimedTNT(
                 this.getLevel().getChunk(this.getFloorX() >> 4, this.getFloorZ() >> 4),
                 nbt, source
